// frontend/src/api/zod/pets/pets.zod.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Pet Store
 * OpenAPI spec version: 2.0
 */
import {
  z as zod
} from 'zod';


export const petsListPetsQueryParams = zod.object({
  "locale": zod.string().optional()
})

export const petsListPetsHeader = zod.object({
  "request-id": zod.string(),
  "client-version": zod.string().optional()
})

export const petsListPetsResponseAgeMin = 0;

export const petsListPetsResponseAgeMax = 100;


export const petsListPetsResponseItem = zod.object({
  "id": zod.number(),
  "name": zod.string().min(1),
  "age": zod.number().min(petsListPetsResponseAgeMin).max(petsListPetsResponseAgeMax),
  "kind": zod.enum(['dog', 'cat', 'fish', 'bird', 'reptile'])
})
export const petsListPetsResponse = zod.array(petsListPetsResponseItem)

export const petsCreatePetQueryParams = zod.object({
  "locale": zod.string().optional()
})

export const petsCreatePetHeader = zod.object({
  "request-id": zod.string(),
  "client-version": zod.string().optional()
})

export const petsCreatePetBodyAgeMin = 0;

export const petsCreatePetBodyAgeMax = 100;


export const petsCreatePetBody = zod.object({
  "id": zod.number(),
  "name": zod.string().min(1),
  "age": zod.number().min(petsCreatePetBodyAgeMin).max(petsCreatePetBodyAgeMax),
  "kind": zod.enum(['dog', 'cat', 'fish', 'bird', 'reptile'])
})

export const petsGetPetParams = zod.object({
  "petId": zod.number()
})

export const petsGetPetQueryParams = zod.object({
  "locale": zod.string().optional()
})

export const petsGetPetHeader = zod.object({
  "request-id": zod.string(),
  "client-version": zod.string().optional()
})

export const petsGetPetResponseAgeMin = 0;

export const petsGetPetResponseAgeMax = 100;


export const petsGetPetResponse = zod.object({
  "id": zod.number(),
  "name": zod.string().min(1),
  "age": zod.number().min(petsGetPetResponseAgeMin).max(petsGetPetResponseAgeMax),
  "kind": zod.enum(['dog', 'cat', 'fish', 'bird', 'reptile'])
})

export const petsUpdatePetParams = zod.object({
  "petId": zod.number()
})

export const petsUpdatePetQueryParams = zod.object({
  "locale": zod.string().optional()
})

export const petsUpdatePetHeader = zod.object({
  "request-id": zod.string(),
  "client-version": zod.string().optional()
})

export const petsUpdatePetBodyAgeMin = 0;

export const petsUpdatePetBodyAgeMax = 100;


export const petsUpdatePetBody = zod.object({
  "id": zod.number(),
  "name": zod.string().min(1),
  "age": zod.number().min(petsUpdatePetBodyAgeMin).max(petsUpdatePetBodyAgeMax),
  "kind": zod.enum(['dog', 'cat', 'fish', 'bird', 'reptile'])
})

export const petsUpdatePetResponseAgeMin = 0;

export const petsUpdatePetResponseAgeMax = 100;


export const petsUpdatePetResponse = zod.object({
  "id": zod.number(),
  "name": zod.string().min(1),
  "age": zod.number().min(petsUpdatePetResponseAgeMin).max(petsUpdatePetResponseAgeMax),
  "kind": zod.enum(['dog', 'cat', 'fish', 'bird', 'reptile'])
})

export const petsDeletePetParams = zod.object({
  "petId": zod.number()
})

export const petsDeletePetQueryParams = zod.object({
  "locale": zod.string().optional()
})

export const petsDeletePetHeader = zod.object({
  "request-id": zod.string(),
  "client-version": zod.string().optional()
})