// frontend/src/api/zod/pets-toys/pets-toys.zod.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Pet Store
 * OpenAPI spec version: 2.0
 */
import {
  z as zod
} from 'zod';


export const toysListToysParams = zod.object({
  "petId": zod.number()
})

export const toysListToysQueryParams = zod.object({
  "locale": zod.string().optional()
})

export const toysListToysHeader = zod.object({
  "request-id": zod.string(),
  "client-version": zod.string().optional()
})

export const toysListToysResponseItem = zod.object({
  "id": zod.number(),
  "name": zod.string()
})
export const toysListToysResponse = zod.array(toysListToysResponseItem)

export const toysCreateToyParams = zod.object({
  "petId": zod.number()
})

export const toysCreateToyQueryParams = zod.object({
  "locale": zod.string().optional()
})

export const toysCreateToyHeader = zod.object({
  "request-id": zod.string(),
  "client-version": zod.string().optional()
})

export const toysCreateToyBody = zod.object({
  "id": zod.number(),
  "name": zod.string()
})

export const toysUpdateToyParams = zod.object({
  "petId": zod.number(),
  "toyId": zod.number()
})

export const toysUpdateToyQueryParams = zod.object({
  "locale": zod.string().optional()
})

export const toysUpdateToyHeader = zod.object({
  "request-id": zod.string(),
  "client-version": zod.string().optional()
})

export const toysUpdateToyBody = zod.object({
  "id": zod.number(),
  "name": zod.string()
})

export const toysUpdateToyResponse = zod.object({
  "id": zod.number(),
  "name": zod.string()
})

export const toysDeleteToyParams = zod.object({
  "petId": zod.number(),
  "toyId": zod.number()
})

export const toysDeleteToyQueryParams = zod.object({
  "locale": zod.string().optional()
})

export const toysDeleteToyHeader = zod.object({
  "request-id": zod.string(),
  "client-version": zod.string().optional()
})